refactor(backend): move get_species_stats logic from database to species service
