+++ backend/src/infra/database/base.rs
@@ -26,7 +26,6 @@ pub trait DatabaseInterface: Send + Sync {
     #[allow(dead_code)]
     async fn execute(&self, query: &str) -> Result<()>;
 
-    async fn get_species_stats(&self) -> Result<Vec<(String, u64)>>;
     async fn get_top_streets(&self, count: u64) -> Result<Vec<(String, u64)>>;
     async fn get_state_stats(&self) -> Result<Vec<(String, u64)>>;
     async fn get_species_mismatch(&self, count: u64, skip: u64) -> Result<Vec<Tree>>;
